ASKAPSDP-2916 - Getting FITS conversion right
The convertToFITS script only does FITS conversion *and header 
updates* when we make CASA images. This means that in FITS mode we 
don't update the headers, which affects downstream tasks like 
continuum validation. This is now fixed - the conversion is still only 
done when needed, but the header updates should always be done.