ASKAPSDP-2929 - Fixing handling of case where num MS != num beams
This moves the check for the number of beams & number of MSs to 
earlier in the pipeline workflow. We now don't terminate the pipeline, 
but give a warning and push on.

This also fixes a bug where the number of beams was only determined 
when the metadata file was being created. Now it is always done (but 
only for the useCLI case).