ASKAPSDP-2591 - Moving away from reporting initial estimate as final fit result in Selavy component catalogue.
This is the first attempt to meet this requirement. We keep the fitted 
result if it meets all fit criteria except chi-squared, but do not 
keep it if it doesn't do this. This results in a catalogue that 
potentially has islands without fitted components. Need to look 
further into improving the fitting for these.