ASKAPSDP-2971 - Enabling a mask in the curvature map
This allows the curvature map used by Selavy to incorporate a mask, 
reflecting both the masked pixels in the input images and also the 
weights cutoff. This entailed changing the implementation of the 
curvature array to a MaskedArray, along with a few interface functions 
in the DistributedImageWriter etc. We also have to pass the Weighter 
into the CurvatureMapCreator to allow this.