ASKAPSDP-2983 - More control over the number of Gaussians in the Selavy fit.
This allows Selavy to reduce the number of Gaussians in a fit in the 
case where the fit does not converge. To do this, we move to iterating 
over a vector list of numGaussians, with resulting changes to the 
relevant loop. We also require finer control over the list of initial 
estimates, so that we can reduce it as needed.