ASKAPSDP-3064 - Fixing the continuum subtraction jobs
This improves the source-finding used in the source-finding jobs in 
the pipeline, changing the Selavy parsets to match what we do in the 
main continuum sourcefinding job. 
This change also streamlines the naming of files, particularly the 
models & component parset that are used - these are now named via a 
utils function which could be called by other scripts.
Also tidying up the continuum source-finding script, appearance-wise 
only.