ASKAPSDP-2912 - Addressing Selavy memory issues with spectral-line sourcefinding
This improves the way Selavy allocates memory when using the 
VariableThresholder class, to try to prevent the spectral-line case 
from failing at large scale. The allocation is done within the loop 
over the planes/spectra in which the variable noise calculations are 
done. The way the writing of the images is also changed, although this 
still has problems when there are more than one subdivision in the 
z-direction.

We also have more control now over the image type for the variable 
threshold output images, and provide functions to get the output image 
names.

There are a few related changes to other files, mostly as a response 
to the imagetype changes and to tidy up some DEBUG output. Formatting 
was done for the files being edited as well.