ASKAPSDP-2912 - Improving source-finding pipeline jobs
This updates the spectral-line sourcefinding pipeline job to work
better with large spectral cubes, taking into account work on the
variable threshold option. We now have new parameters to control the
overlap between subsections, and we turn off the writing of the
variable-threshold images in the spectral-line case. The filenames for 
those files should now be consistent with the image type.