ASKAPSDP-3163 - Fixes to handling of FITS suffixes
This removes the ADD_FITS_SUFFIX parameter from the 
defineArtifactsScript and those who call it. We instead fall back to 
using the IMAGETYPE variables in line with the rest of the pipeline.

Also a couple of minor fixes to applyBandpass1934 & casda scripts to 
tidy up the code and add a backslash.