ASKAPSDP-3169 - Improvements to the distributed writing of beam logs
This adds a gather function to the BeamLogger class, whereby the beam 
lists are gathered together on a nominated rank from all workers.

This is utilised by the ContinuumWorker in the askap_imager, for the 
case of multiple writers on a single cube. The writing of the beam log 
is now done only by the creator rank. A new method was added to the 
CubeComms class to get the list of creator ranks.