ASKAP-3169 - Finalising distributed beam writing.
This fixes the beam writing in the parallel/distributed case. The 
BeamLogger class now does the gathering correctly (taking into proper 
account the rank of each worker), and we remove the calls to 
logBeamInfo from ContinuumMaster, as this is not needed - the gather 
function now has an option to ignore the master.