ASKAPSDP-3235 - Fix for the Selavy WCS error bug
The fix involves changing the way the new initial estimates are set up 
when we want to add a new component to the fit. We were using the fit 
results, when we should have been using the original list of initial 
estimates. This was causing the inclusion of spurious components, 
which caused errors when they were parameterised. 
This fix seems to do the correct thing, and initial tests indicate the 
WCS error goes away.