ASKAPSDP-3320 - Further scripting fixes.
This adds more robustness to the defineArtifactsScript.sh script, 
adding in additional quotes around filenames and improving the tarring 
up of the CMPL tarfile - to do filenames with spaces, we now do a 
series of "tar rvf" commands for each individual file, rather than 
trying to preserve them in a list.

Also fixing loops over arrays, so that we get the full filename at 
each iteration and not just a part of it. This is the case for 
convertToFITS.sh as well.