ASKAPSDP-3560 - New multi-job selfcal option for the pipeline.
This adds a new script that launches multiple jobs for the continuum 
self-cal. There will be one imaging job and one "calibration" job for 
each selfcal loop (the 0-th loop only does imaging, as before). The 
calibration job contains the selavy, cmodel and ccalibrator tasks, all 
set up to fit on a single node. The jobs are chained together with 
dependencies - the resulting dependency string is passed on to 
subsequent tasks (with a minor change to addDep() to make this work).

This mode is triggered via the new input parameter MULTI_JOB_SELFCAL, 
turned off by default.