ASKAPSDP-3289 - Adding a switch to control use of the fat node mode in continuum imaging.
This adds the pipeline parameter FAT_NODE_CONT_IMG, that turns on the
use of the fat node mode. I had to change the srun calls to have a
parameterised command-line option that either uses the node list or
sets ntasks-per-node.

This has been set for all three continuum-imaging jobs.

Updating the user documentation as well