Defining << overload within the askap namespace
I still don't understand exactly how the previous overload in AskapUtil
and casacore's worked correctly together. casacore's is defined within
the casacore namespace, so it's clearly meant to be used only within
that namespace, or with objects of that namespace. I think it would also
become visible when doing things like "using namespace casa(core)",
which explains some of its uses. In some other places there is no such
wide name imports, and code was indeed failing to see the operator.

Putting back the overload as it was defined (i.e., in the std::
namespace) is again too much, as casacore's and std's implementations
collide and the compiler doesn't know which one to use.

I think a better way of solving this is to declare the overload within
the askap namespace. This will make it visible again to all the askap
software, will not collide with casacore's definition, but its
implementation actually still reuses casacore's implementation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>